# docs/security/openapi.yaml
openapi: 3.0.3
info:
  title: Civis CRM/ERP API
  version: 0.1.0
  description: |
    Generated from the current Next.js App Router endpoints under `app/api/**/route.ts`.
    This spec is intended for Swagger/SwaggerHub documentation and VAPT planning.
servers:
  - url: https://ik-crm-erp.onrender.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    nextAuthSession:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie (development/non-secure cookie name).
    nextAuthSecureSession:
      type: apiKey
      in: cookie
      name: __Secure-next-auth.session-token
      description: NextAuth session cookie (secure cookie name).
  schemas:
...
    ContactInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        company: { type: string }
        status: { type: string, enum: [LEAD, PROSPECT, CUSTOMER] }
        revenue: { type: integer }
        lastContact: { type: string, format: date-time }
        tags: { type: string }
        ownerId: { type: string }
        notes: { type: string }
        customFields:
          type: object
          additionalProperties: true
      required: [name, email]
...
security:
  - nextAuthSession: []
  - nextAuthSecureSession: []
paths:
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List organization users
      responses:
        '200': { description: User list }
        '403': { description: Admin required }
    patch:
      tags: [Admin]
      summary: Update a user's role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string }
                role: { type: string, enum: [USER, ADMIN, SUPER_ADMIN] }
              required: [id, role]
      responses:
        '200': { description: User updated }
        '403': { description: Admin/Super-admin required }
...
  /api/crm/contacts:
    get:
      tags: [CRM]
      summary: List contacts
      responses:
        '200': { description: Contacts list }
    post:
      tags: [CRM]
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContactInput' }
      responses:
        '200': { description: Contact created }
        '400': { description: Validation error }
    patch:
      tags: [CRM]
      summary: Update contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/ContactInput'
                - type: object
                  properties:
                    id: { type: string }
                  required: [id]
      responses:
        '200': { description: Contact updated }
    delete:
      tags: [CRM]
      summary: Delete contact
      parameters:
        - name: id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: Contact deleted }
...